third_party/WebKit/LayoutTests/overflow/overflow-basic-001.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#container {
  overflow: auto;
  border: 1px solid black;
  width: 200px;
  height: 100px;
}
#target {
  width: 300px;
  height: 200px;
  background-color: rgba(0, 255, 0, 0.3);
}
</style>
<p>Overflow test.</p>
<div id="container">
  <div id="target"></div>
</div>
<script>

var container = document.querySelector('#container');
var target = document.querySelector('#target');

test(function() {
  assert_equals(container.scrollWidth, target.offsetWidth, "width");
  assert_equals(container.scrollHeight, target.offsetHeight, "height");
}, "simple overflow");

test(function() {
  target.style.height = "400px";
  target.style.width = "400px";
  assert_equals(container.scrollWidth, target.offsetWidth, "width");
  assert_equals(container.scrollHeight, target.offsetHeight, "height");
}, "simple overflow, after target resize");
</script>
